Add action to create a check run
Currently, this action only creates a commit status, not an actual check run.
The shortcoming is that it does not allow us to set the status to "in-progress".
That is why you will see status as pending, but the description will say
Enterprise tests are running".

To fix this discrepancy, we need to create a check rather than a commit status.
This gives us more control over the status of the check.
More documentation here: https://docs.github.com/en/rest/checks/runs?apiVersion=2022-11-28#create-a-check-run.

For now, I created a new action so as to not interfere with any existing action
and not break anything in CI. After this, I want to slowly migrate the existing
CI workflows which use commit statuses to this new action.